Refactor WebSocket imports and restructure file paths for improved organization
